En del ändringar i flera klasser, har även skapat HighScoreManager

+++ src/ghzclicker/ScoreComparator.java
@@ -8,8 +8,8 @@ import java.util.Comparator;
  */
 public class ScoreComparator implements Comparator<Score> {
 	public int compare(Score score1, Score score2){
-		int sc1 = score1.getScore();
-		int sc2 = score2.getScore();
+		double sc1 = score1.getScore();
+		double sc2 = score2.getScore();
 		
 		if(sc1 > sc2){
 			return -1;
